syntax: methodify the lexer
