+++ src/libsyntax/print/pprust.rs
@@ -20,7 +20,8 @@ use codemap;
 use diagnostic;
 use parse::classify::expr_is_simple_block;
 use parse::token::IdentInterner;
-use parse::{comments, token};
+use parse::token;
+use parse::lexer::comments;
 use parse;
 use print::pp::{break_offset, word, space, zerobreak, hardbreak};
 use print::pp::{Breaks, Consistent, Inconsistent, eof};
